backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
langchain
langchain-community
langchain-groq
chromadb
gitpython
tiktoken
chardet